Added DisplayMask class to manage frame decorations etc
